requirements.txt
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
requests>=2.28.0
lxml>=4.9.0
python-dateutil>=2.8.0 